+++ src/Service/CustomField.php
@@ -2,24 +2,22 @@
 
 namespace Nickcheek\Brightree\Service;
 
+use Nickcheek\Brightree\Brightree;
 use SoapClient;
 
-class CustomField
+class CustomField extends Brightree
 {
-	protected $custom;
 	protected $custom_options;
 	
 	public function __construct()
 	{
-		DEFINE("BASE", dirname( __FILE__ ) ."/" );
-		$config = include(BASE . '../config/config.php');
-		$this->custom = $config->service['custom'];
-		$this->custom_options = array('login' => $config->user['name'],'password' => $config->user['pass'],'uri' => $this->custom,'location' => $this->custom,'trace' => 1);
+		parent::__construct();
+		$this->custom_options = array('login' => $this->config->user['name'],'password' => $this->config->user['pass'],'uri' => $this->config->service['custom'],'location' => $this->config->service['custom'],'trace' => 1);
 	}
 	
 	public function apiCall($call,$query)
     {
-        $client     = new SoapClient( $this->custom .'?singleWsdl', $this->custom_options);
+        $client     = new SoapClient( $this->config->service['custom'] .'?singleWsdl', $this->custom_options);
         $response 	= $client->$call($query);
         return $response;
     }